requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.40
aiosqlite==0.21.0
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
starlette==0.46.2
jinja2==3.1.6
pdfkit==1.0.0
weasyprint==65.1
requests==2.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytest==8.3.3
pytest-asyncio==0.25.0
httpx==0.28.1